+++ internal/bins/bins.go
@@ -16,6 +16,7 @@ import (
 	"github.com/blang/semver/v4"
 	"github.com/cupcakearmy/autorestic/internal"
 	"github.com/cupcakearmy/autorestic/internal/colors"
+	"github.com/cupcakearmy/autorestic/internal/flags"
 )
 
 const INSTALL_PATH = "/usr/local/bin"
@@ -128,10 +129,9 @@ func InstallRestic() error {
 }
 
 func upgradeRestic() error {
-	_, out, err := internal.ExecuteCommand(internal.ExecuteOptions{
-		Command: "restic",
+	_, _, err := internal.ExecuteCommand(internal.ExecuteOptions{
+		Command: flags.RESTIC_BIN,
 	}, "self-update")
-	colors.Faint.Println(out)
 	return err
 }